Moved to C++20. More fixes for async directory/files tree loading in the script profiles window.
Not yet thoroughly tested. Working on a stable release.

+++ src/uoclientfiles/uoanimuop.h
@@ -24,8 +24,8 @@ class UOAnimUOP
     struct UOPAnimationData
     {
         int animFileIdx;        // AnimationFrame(x).uop
-        size_t blockIdx;
-        size_t fileIdx;
+        unsigned int blockIdx;
+        unsigned int fileIdx;
         unsigned long long hash;
     };
